requirements.txt
Django>=5.0,<6.1
djangorestframework>=3.16
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
psycopg2-binary>=2.9   # Only needed if using PostgreSQL
